+++ modules/core/ui/map/js/farmOS.map.behaviors.asset_type_layers.js
@@ -1,4 +1,4 @@
-(function () {
+(function (drupalSettings) {
   farmOS.map.behaviors.asset_type_layers = {
     attach: function (instance) {
 
@@ -110,4 +110,4 @@
       });
     }
   };
-}());
+}(drupalSettings));